Bluetooth: Move CTS implementation to subsys/bluetooth/lib
This moves GATT Current Time Service implementation under
subsys/bluetooth/lib and integrate with Kbuild to avoid build
problems.

Fixes #1166
