+++ samples/bluetooth/peripheral/prj.conf
@@ -8,3 +8,4 @@ CONFIG_BT_PRIVACY=y
 CONFIG_BT_DEVICE_NAME="Test peripheral"
 CONFIG_BT_DEVICE_APPEARANCE=833
 CONFIG_BT_BAS=y
+CONFIG_BT_CTS=y
