+++ samples/bluetooth/peripheral/src/main.c
@@ -20,10 +20,10 @@
 #include <bluetooth/uuid.h>
 #include <bluetooth/gatt.h>
 #include <bluetooth/bas.h>
+#include <bluetooth/cts.h>
 
 #include <gatt/hrs.h>
 #include <gatt/dis.h>
-#include <gatt/cts.h>
 
 #define DEVICE_NAME		CONFIG_BT_DEVICE_NAME
 #define DEVICE_NAME_LEN		(sizeof(DEVICE_NAME) - 1)
