+++ subsys/bluetooth/lib/Makefile
@@ -1,1 +1,2 @@
 obj-$(CONFIG_BT_BAS) += bas.o
+obj-$(CONFIG_BT_CTS) += cts.o
